Convert "trigger anywhere" from botwide setting to per-fact
This requires a schema update, so I also added some very basic schema
versioning stuff. v0 = pre-versioning, v1 = pre-this-update, v2 =
supports "match anywhere" functionality.

The global option still exists, and this just controls how new facts are
created. If TRIGGER_ANYWHERE is in the environment as 1, then new facts
will be created with MATCH_ANYWHERE set. If it is set to 0 or not set,
then new facts will NOT match anywhere. Currently the "recommended"
"procedure" for having facts that buck the global trend, would be for
the admin to use the sqlite3 command-line tool to manually alter the
fact in the database.